[RPP] Refactor the `draw` function into multiple functions for simplicity
The `draw` function in the FlameChart class draws the perf panel flame
chart into the canvas. Due to the complexity of the data being drawn and
little use of helper functions, the function was too long and hard to
understand.

This CL splits the function into multiple helper functions, aiming to
improve readability. Each function is also documented with a short
description.

Bug: 1386091
Change-Id: I24b5325b4d1ed133d4ae81bd5c4c54149e59fc8b