Move logpoint detection from view to SDK layer
We want to render an icon for console messages that originate from
logpoints/conditional breakpoints.

This means we need to identify such console messages in more than
one place, not just for anchor rendering.

This CL moves the corresponding code into the SDK ConsoleMessage
class, where it probably should have been in the first place.

R=jarin@chromium.org

Bug: 1027458
Change-Id: I1a443e52ae0aecdf5368bef124c65c629abad70e