Add tests to TimelineVisibleEvents
Adds some tests to the TimelineModelFilter code which previously had
none. This CL is not exhaustive, there are more tests to write but I
will do them in follow-up CLs.

I also made the base class an abstract class, to clarify the intent that
it is only ever subclassed and never created directly.

Bug: 1414422
Change-Id: I1f1373a4ccb070b4024368dcd04c9467c425904f