[sources] Try remove entire project on execution context destroyed
If an execution context gets destroyed, this patch tries to figure
out if an entire project gets removed. If yes, we will issue
ProjectRemoved event rather than UISourceCodeRemoved for each
source code in the project. This helps to throttle UI updates
when a whole project gets removed (e.g., when a frame gets removed).

Bug: chromium:1409025
Change-Id: Iff1c870761b5a53f3842fa7f0cf1e58df1692319