[CSSHints] Fix incorrect checks for the presence of parent styles
For pseudo elements, we currently cannot find the correct parent styles so the correct behaviour in this case is to show no hints.

The bug was the check for a null instead of an undefined. Unfortunately, TS didn't help us to catch this. I have also refactored all other instances and removed the #isRuleValid helpers because it is not completely clear what is a rule and when it is valid.

Fixed: 1414631
Change-Id: Ib047bddbaf54c1112878b374bcde557610fb2943