[console] Replace UI.Icon.Icon with Lit component for message icons
We want to add icons for logpoint / conditional breakpoint messages
and we don't want to add those icons to the legacy sprite sheet.
Since that would mean mixing `UI.Icon.Icon` and `IconButton.Icon.Icon`
it's cleaner to update the existing icon code first. Luckily the
warning and error icon have already been extracted from the sprite
sheet and are available.

This CL is visually a no-op.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: I72bf034b725e49daca8b4d4b6a59ba774312877b