[bindings] Avoid `ResourceMapping.getMappedLines()` for now.
This undoes part of the change from https://crrev.com/c/4161371,
effectively ignoring `getMappedLines()` for the `ResourceMapping` for
now. The reason here is that for the `ResourceMapping`, scripts can
appear at any later point in time, and we have no way currently to
thread through a signal to the UI to recompute the mapped lines.

To get back to a working state, we'll simply disable this new
functionality for now until we have a better solution.

Bug: chromium:1185770
Change-Id: I37d66b2e4ff2a61a8abc5b95d734394aac8fb71d
Fixed: chromium:1411431