Rework TimelineModel tests to have no globals
The tests were all using the same global timelineModel which was causing
me confusion when adding new tests. I have now updated it so each test
uses a method to create an instance of the model with events, which is
much clearer. This will make new tests (follow up CLs) easier to add.

Bug: 1414422
Change-Id: I9f52559d844b36d7a0d62029ebc5c4e76c9df451