Add all related uiSourceCodes when setting a breakpoint
In order to make sure that we can successfully resolve
uiLocations for all targets with the same script, we need to
include all their uiSourceCodes when setting a breakpoint.

Currently, only one uiSourceCode is added on e.g. user interaction. The remaining targets that don't have their uiSourceCodes included
will use a fallback, that can be racy and give different results.
With this patch, we search for all related uiSourceCodes and
explicitly set the breakpoint on each of them.

Bug: 1415258
Change-Id: I8e1becf3efcaf159aa9631c817c17c77974a435d