Add all related uiSourceCodes when setting a breakpoint
TLDR: In order to make sure that we can successfully resolve
uiLocations for all targets that run the same script, we need to
include all their uiSourceCodes when setting a breakpoint.

At the moment, when opening a uiSourceCode and setting a breakpoint
in the UI, only one single uiSourceCode is added to the `Breakpoint`
class. This uiSourceCode will be used to resolve the raw location
of the breakpoint. The raw location is then sent to the backend.

If multiple targets run the same script, DevTools will try
to set a breakpoint for all of them. The problem arises when we
resolve the raw location: while we manage to resolve it for the one
target whose uiSourceCode was chosen when clicking on the UI, we
are not able to correctly resolve the raw location for other targets,
as their uiSourceCodes was not added in the `Breakpoint` class.

In this case we have a fallback case: using whatever raw location
another target has already set. This can turn out correct, or wrong.
Instead, we want to make sure that all related uiSourceCodes are added.


Bug: 1415258
Change-Id: I8e1becf3efcaf159aa9631c817c17c77974a435d