Deflake breakpoints added to workers tests
The reason for the flakiness is at `addBreakpointForLine`; since there is no visiual indicator that the setting of the breakpoint is complete the script that is supposed to trigger the breakpoint is ran before it is created, resulting in it timing out at await `waitFor(RESUME_BUTTON);`. This was apparent because adding a wait after this line significantly reduced the flake. I tried to fire an event and listen to it in the test but I could not pinpoint where in the code the breakpoint creation is complete, at the end the cheapest solution was to retry running the script until we successfully pause. Please find the below stress runs to confirm this:

https://ci.chromium.org/ui/p/devtools-frontend/builders/try/e2e_stressor_mac/b8789172354377813809/overview
https://ci.chromium.org/ui/p/devtools-frontend/builders/try/e2e_stressor_win64/b8789172354377813793/overview
https://ci.chromium.org/ui/p/devtools-frontend/builders/try/e2e_stressor_linux/b8789172354377813825/overview

Bug: 1368493
Cq-Include-Trybots: luci.devtools-frontend.try:devtools_frontend_mac_rel
Change-Id: I98eebdd497e4f793847e52cd1e16b54aaa4056f5