Add ConstructedEvent and PayloadEvent to TracingModel
This is the first step in differentiating between events that map 1-1
onto a trace event payload and those that we construct.

Also includes a drive-by fix to the ESLint custom naming rule to fix an
error when the class it is trying to lint does not have a constructor
and does not call super(). A further fix for the fact that this custom
events rule flags false positives in this file is implemented in
crrev.com/c/4259791, at which point we will not have to disable the rule
in this file.

Bug: 1416836
Change-Id: Ia943fa95aeed74b6dc5e51d240a92cc945d02a89