[sources] Further simplify resource origin (also UI-wise).
This removes the unnecessary special casing for the language plugins,
where we generated a different toolbar string ("provided via debug
info") instead of the regular one ("source mapped from"), which was a
weird special case, since everywhere else in the UI we also use the term
"source map" (for example when hovering over tree items in the Page
tree).

We also update the UI to bring it on par with Firefox DevTools, which
shows a simple and elegant "From FOO" in the corner. If there are
multiple origins, we simply enumerate them all (separated by commata).

Bug: chromium:1416912
Change-Id: I59bce85068ab7db6cbcf9bbcb1c014e475a56bcc
Demos: https://imgur.com/BFJZCtk.png & https://imgur.com/fUblSvY.png