Fix false positive in enforce_custom_event_names lint rule
The issue here was that in SDK we have an actual class called Event,
which was messing with the rule which assumed that any extending of
Event was extending the built in HTML Event class. The rule is now
updated to detect locally declared Event classes, and not run the rule
in this instance.

Bug: none
Change-Id: Ib90054728566cb55d367a241130d70b81aeb1e79