[RPP] Throttle the name resolving pipeline using a timeout
Resolving funcion names of a CPU profile using source maps used to be
triggered every time a source map was attached to DevTools. This is
problematic because after names are resolved, the flame chart is
redrawn, meaning that multiple source maps would trigger multiple
renders.

This CL adds a timeout that triggers the reload sometime after a
source map is attached. Until names are resolved, no other
source-map-attached event can trigger a new timeout. This way we
allow for other source maps to be parsed before attempting a name
resolving. If a new source map is then parsed, the cycle starts over.

Fixed: 1416594


Fix ESLint custom names not respecting custom Event class

This ESLint rule assumes that if a class extends an Event class, it is
extending the built in class and applies checks accordingly. However, if
we are in code that has a local Event class defined, then we should not
apply this check.

Bug: none
Change-Id: I1ce94299074c44812fd2845fe7e487a2c14b140c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4259791
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>


Move navigator view throttler to Common

This way it can be more easily used in other parts of DT

Bug: none
Change-Id: I9afb28adc6cbc5361a7b88795a8c6052e2c3d1be