[RPP] Throttle the name resolving pipeline using a timeout
Resolving funcion names of a CPU profile using source maps used to be
triggered every time a source map was attached to DevTools. This is
problematic because after names are resolved, the flame chart is
redrawn, meaning that multiple source maps would trigger multiple
renders.

This CL adds a timeout that triggers the reload sometime after a
source map is attached. Until names are resolved, no other
source-map-attached event can trigger a new timeout. This way we
allow for other source maps to be parsed before attempting a name
resolving. If a new source map is then parsed, the cycle starts over.

Fixed: 1416594
Change-Id: I9afb28adc6cbc5361a7b88795a8c6052e2c3d1be