Do not duplicate breakpoints on file system removal
When a file-system-network binding gets removed, we copy breakpoints
from the network source code to the file system source code so that
the breakpoints do not disappear.

However, this is problematic if the binding was removed because
the file system was removed from the workspace. In that case,
we would copy the breakpoint into a source code that is being
removed.

This CL detects copying breakpoints to removed source code and
instead of creating a live breakpoint, it simply adds an entry to
the breakpoint storage.

Bug: chromium:1416164
Change-Id: I26f073fa76a0799e039d7ff5553f9b9597864f3c