Reveal breakpoint edit dialog for structured stack trace links
This CL changes the structured stack trace rendering to reveal the
breakpoint edit dialog should the console message/error originate
from a breakpoint condition.

The heuristic is very simple: Check if the parent frame is one of the
special marker scripts, if so, check if the current source location
has a breakpoint.

Note that the logic is slightly less sophisticated as for the
console message anchor, but this only matters if the user starts
leaking closures from breakpoint conditions.

R=bmeurer@chromium.org

Bug: 1027458
Change-Id: I259b8a7614635598e6a7628cb94b8c0dbfbf74dc