Build helper for mock payload events
This CL introduces a better helper for constructing mocked PayloadEvents
which takes in items for the payload, rather than a helper that takes in
some options that do not necessarily map directly to the keys found in
actual trace events.

This enables us to more accurately mock events. I have now removed the
other helper method and swapped all usages out with this new method, and
added documentation.

Bug: 1414422,1416836
Change-Id: Icda5b26b13306bf8a12cfeaf0613aff8a6383891