[console] Make ArrowDown select first suggestion.
This addresses an annoying inconsistency in the conservative
autocompletion, where despite not having selected anything
from the list effectively, pressing ArrowDown / Ctrl+n takes
you to the second selection in the list.

Now with this change, we first check whether completion state
is conservative and if so, forcibly select the first item.

Bug: chromium:1276960
Change-Id: I07e3f64394d1c50351329dbc5a8e74d54c1b1b9b