Add support to clear ActionRegistry and ActionRegistration in one call
It's probable that when ActionRegistry.removeInstance is called, one
would also like to remove all the registered actions at that point. This
is specially probable in unit-test scenarios. For this reason, a new
`reset` function is added to action registry that achieves both things.

As a drive-by fix, the actions added in the action_registration test are
removed so that they don't pollyte other tests.

Bug: none
Change-Id: Iaf49c53e61b3f9159a48b369693e5a40fcef44f7