[RPP] Execute both engines in parallel and pass both results to UI
A call to execute the new engine is done in parallel to
PerformanceModel.setTracingModel. The latter contains the main parsing
logic of the OPP, in terms of the time it takes to be executed This is
where the trace data is transformed to the representation that is used
by the subsequent UI layer.

Fortunately performanceModel.setTracing model doesn't do any rendering
rendering, which gives us the opportunity to pass the results of
invoking it along with the results of the new engine to the layer that
does the rendering simultaneously.

The first component in the UI layer (after parsing) is the
TimelineFlameChartDataProvider (data provider). The result of the
parsing is passed to it, which from this CL includes both the old
PerformanceModel and the new engine's trace parsed data. This CL also
does some renaming in the data provider to distinguish between the data
that will be replaced (legacy) and the data of the new trace engine.

Bug: 1416533
Change-Id: I4161609db9eb6186cc6a788f9ea60e886502f13f