Add support for new engine's parsed data in TimelineHistoryManager
This CL does the minimum changes to the TimelineHistoryManager so that
it works properly using the data coming from the new engine.

The history manager still relies on a lot of features the legacy
PerformanceModel provides and that we should implement in the new
engine in the future to remove the legacy components completely.

Bug: 1416533
Change-Id: Ib9d5b451569d9f6e83ab90b7fabfb138fbf317c4