Remove auto resizing behavior from CSS plugin
This returns the behavior of CSS tooltips in
sources panel to be prior to this change:
https://github.com/codemirror/view/commit/b8cec629df04969a64f2d2c1e2b85b8e1ca8de00
which caused color picker to have a weird shape.

We could have added an `overflow: auto` to make
sure that the color picker is correctly rendered with
a defined height but it causes the widget to be
flickering when it is interacted with and re-rendered.

Fixed: 1412976
Change-Id: I256bdbaebc5acde79a26a28ae90efc46bc56b9bb