Introduce 'TextEditorHistory' class
We extract the `moveHistory` logic into a small helper class. It syncs
moving through the history with a CodeMirror instance.

The bulk of the logic is about cursor positioning: Whether to move or
not move to the next history entry depending on where the cursor
is currently positioned.

The CL adds unit tests for all of these to better document the
existing behavior.

R=kimanh@chromium.org

Bug: 1027458
Change-Id: Ie842d7b57c580c50ad956778dffea5c41217d4d2