[RPP] Add new TrackAppender layer to timeline flamechart architecture
This adds support from another source of data in the
TimelineFlameChartDataProvider, as proposed in the doc to plug the new
engine to the RPP main flame chart [1]. This way, the flame chart is
drawn with data coming from the existing code paths and from the new
CompatibilityTracksAppender.

At the moment, no data is supplied from the new source and no data is
removed from the old source. This only the mere boilerplate to start
migrating tracks to be added progressively using the new track appender
system.

[1] https://docs.google.com/document/d/1kQlWn5c64YBi9SdXQYPK9zDXTNLKIs5VyPIapB7tYAM/edit?resourcekey=0-4_RFds4ZZ0iTTe310C0TgA#heading=h.7nki9mck5t64

Bug: 1416533
Change-Id: I8eb0830e649968fc572eb8656304d253e2fc955b