Make heap snapshot diffs more accurate
When comparing two heap snapshots, devtools will sometimes report that
an object was deleted and created, despite the object having the same ID
in both snapshots. This can happen because calculateDiffForClass expects
both lists of nodes to be sorted by ID, but the lists aren't always
sorted. It appears that getAggregatesByClassName(true, 'allObjects')
should produce sorted results due to passing true as the first argument,
but it is sorting the wrong collection.

Bug: chromium:1286500
Change-Id: I73e68179cdfee1c6beacf04ecef5824e6245f8ac