[sources] Work around for webpack-based Vue2 set ups.
This is a sad but necessary work-around. Before we had proper *.vue
support in DevTools, we'd highlight all *.vue files as JavaScript (which
wasn't really working as expected, but worked well enough). As a
consequence of this, in Vue2 (which is based on webpack), all the weird
webpack artifacts that aren't actually Vue SFCs, but instead some
generated JavaScript, which show up as `<file>.vue?<hash>` served by
webpack-dev-server, would be highlighted properly as JavaScript, simply
because there was never support for Vue SFCs.

While this work-around doesn't fix the fundamental problem (the fact
that the webpack-dev-server shows generated code in the first place), it
does mitigate the pain of developers, since the JavaScript that webpack
presents is at least highlighted as such.

We do this by sniffing the contents of *.vue files in the SourceFrame,
and when the content starts with a `<` we assume that it's a proper Vue
SFC file, otherwise we assume it's some generated JavaScript.

This partly undoes https://crrev.com/c/4022706 in that it brings back
content sniffing to the language detection in the SourceFrame.

Fixed: chromium:1416562
Bug: chromium:1320204
Change-Id: I4e8dbe267735ca560340e773193f95cb760279e5