Add console stub to trace worker
Because the trace model runs in a worker, we cannot console.log from it
when running Karma in the terminal and see the output, because Karma
does not (and cannot) fake the console in the worker execution context.

Instead, we manually mock part of the console API within the worker, so
we can console.log/warn/error and see the output in the terminal. The
code that runs the worker listens to these messages and calls console in
its main context, which Karma is hooked into.

I considered making this a generic solution that can be easily
implemented into every worker, but given we have not needed this before
today for this one worker, I think for now we should keep it local and
only wrap it in a helper method if required in the future.

Bug: none
Change-Id: Iab05f6a9a729960dd5f4d53219cc0d9b4f65cb98