Add unit test for generic trace support to new engine
This CL adds a basic test that we can load a generic trace - not one
generated from DevTools - without erroring. Note that there may be more
we need to do here in the long run to support these traces better, but
this test will ensure at the least that we do not regress to the point
where we cannot load them.

Bug: 1417612
Change-Id: I27a3755a82b4d39df68fad804646cdbcfd99ec20