[RPP] Add a new entry type to flame chart data provider
The new entry type is reserved for those events that are appended to the
data provider by the new "TrackAppenders". This CL includes the handling
of the data provider callbacks used by the canvas renderer (FlameChart)
to get extra features for an event (f.e. its color, title, etc). The
handling consists of invoking the compatibility layer to handle the
call appropriately.

Gradually migrating tracks to the new approach means there will be no
other type of entry, since all events will be handled by the
compatibility layer. At that point, this layer won't be a
"compatibility" layer, but just a layer. Most likely, at that point,
we should be able to remove the TimelineFlameChartDataProvider
completely.

Bug: 1416533
Change-Id: I0cba4b18e9eb4668d826a4ea5f141938c51b67d9