[RPP] Add support for performance.mark events in UserTimingsHandler
Previously the UserTimingsHandler only supported events coming from the
performance.measure API.

So that we can use the new engine in the Timings track, this CL adds
support for events from the performance.mark api. To gather them events
all filtered by the blink.user_timing category and then filtered by the
known names of events injected by Chrome. This emulates the parsing
behaviour of the OPP.

A drive-by change is included where all the async phases are added to
our Phase enum as well as a type and type helper for async events.

Bug: 1409044