Fix errors on traces with FLEDGE
When looking into the support for OPP for FLEDGE frames (introduced in
crrev.com/c/3686811) I discovered that the new trace engine errors on
them. This is because in the test trace, there seem to be two navigation
events for the same thread but different process ID, which breaks an
assumption we had. I am going to explore this further and chat to people
who can hopefully help clarify the behaviour here, but in the mean time
let's fix it so it does not error.

Bug: none
Change-Id: Ie8a770855d24c9e7d59d4215eaf04dcc1d0cea1a