Use new engine isLCPEvent helper
This CL takes advantage of the fact that we now store raw payloads and
can consequently get them and pass them into helpers we define
downstream.

It will get a little tricky because the helpeers downstream do not care
about frames and main frames; in Insights we leave the UI layer to
decide which events to show or not show, but in OPP that selection is
done in the data layer, so we still have to keep that logic around.
However this still will enable us to remove some logic from OPP that we
can defer to the new engine.

Bug: 1416836
Change-Id: Ifeab9092b779237ffb35807b70c324c6207e0a77