[RPP] Add support for marker events in PageLoadMetricsHandler
Marker events are page load events that have no duration and that appear
in the Timings track. For example, largest contentful paint,
MarkDomContent DomContentLoad, etc. Although the PageLoadMetricsHandler
buffered some of these, they were not exported and not all marker event
types were buffered (like the Load event).

So that we can use the new engine in the Timings track, this CL adds
support for marker events in the handler by buffering and exporting all
the marker events in a trace.

Bug: 1409044
Change-Id: Ifb0a231f1818f02357af21797d24cca733b70892