[HeaderOverrides] Get request by id instead of by URL when marking as overridden
When headers are being overridden, the NetworkRequest is marked as
'intercepted'. If there are multiple requests with the same URL,
finding the request by URL causes problems since URL is not unique.
The consequence is that only one of the requests is marked as
'intercepted'.
With this CL the requests are referred to by their ID instead, which is unique, and therefore all requests are correctly marked as
'intercepted'.

Fixed: 1420271
Change-Id: I5f9c445a1e54958bcb6ce1e0a76957448fd81ffd