[RPP] Use the new engine in the Timings track
The first track appender is introduced to add the Timings track into the flame chart data (TimingsTrackAppender). This appender consists of the bits previously in the TimelineFlameChartDataProvider that were in charge of appending the data. All this parts are moved to a common place to handle the details of the track together in the same module, but separate from other tracks. The source of data used by the track appender is the output of the new engine

In the future further test coverage will be added using dedicated screenshot tests.

This has to be landed once the new engine has been reenabled

Bug: 1409044
Change-Id: I007d5ffd93e310cba26546fa09829e02e3cc581c