Automatically reset trace handlers in tests
This CL automatically resets trace handlers between tests so we know
that no data is leaking into other tests. I then found some tests that
needed fixing, and I also updated the network handler to have an
explicit initialize step, which brings it inline with the other
handlers and makes it easier to test.

Bug: none
Change-Id: I0f1c055658b8639274406ee392bfb773938957f3