[Merge into M112] Remove execution of new trace engine
This is the backmerge of crrev.com/c/4295978 into the M112 branch.

Original CL description:
It has implications on the time taken for DevTools to process a trace,
and right now on ToT we are not using the data from the parsing. Let's
remove it for now, and bring it back when we are using the data.

Bug: 1420240
Change-Id: I6f0932715534783fbddbf9b9101a718109d2be1a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/4295978
Reviewed-by: Andres Olivares <andoli@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
(cherry picked from commit 274b8707ee74749401d0d555ed0cea3c43ffb690)