Show memory icon next to all viable types
With this CL, memory icons (to open the memory inspector) will be
shown next to all viable types (including data views, typed arrays,
which were previously omitted), in the scope view, in the sources view,
and also in console messages.

To prevent the memory icon from cluttering the view when expanding
an object, we omit it next to any synthetic properties, e.g. on
typed arrays that are shown as internal properties.

Bug: 1415192
Change-Id: I1568872dfc3dabb50e41f8c0c06ff54e5bdeee7b