CRbug 1422353: Performance panel can be broken
A timing issue with the interval for updating the Loading progress bar
can cause an exception to occur, blocking the Performance panel from
progressing beyond "Loading profile...". This is because when the
TypeScript port took place, the `time` field was marked with a non-null
type assertion, but this type assertion is not accurate, as the field
might not be initialized depending on the parameters passed into the
StatusPane constructor.

This fix removes the non-null type assertion, correctly types the field,
and fixes the compiler error that would have been generated if the field
had the correct type previously.

Bug: 1422353
Change-Id: Ia0f4f06c8b616adb95793722ba5a67c657024ea6