[RPP] Prepare main flame chart code base for screenshot testing
Adds a helper that renders a flamechart using a trace file and a set of
tracks passed as parameters. For this purpose a method is added to the
tracks appender that allows to select the tracks that are visible by
passing a set of track names.

The data provider is also changed to include a method that builds the
flame chart data using only the new track appenders. Also, a method to
access to the internal compatibility tracks appender is added so that
tests can have a reference to it, for example, to set the desired
visible tracks.

Finally, added some JSDocs to relevant methods.

Bug: 1409044
Change-Id: I2409d544bc1eb7f7659c063e318cfe17e1c9665c