Partially parse animation shorthand and render keyframe links
I was thinking parsing `animation` shorthand to be an overkill
in crrev.com/c/4300470 however it felt hacky to do otherwise and I've
finally opted in to partially parsing it.

Created a `CSSAnimationModel` that tokenizes the given shorthand
value and finds the parts in it that should be rendered as text,
easing function or animation name.

I did not split the text part into other longhand properties since we
render them as text anyways. If we feel the need for this (i.e. a new
swatch for some longhand property), we can do that on top of this patch.

Here is the screenshot for animation shorthand to keyframe links:
https://bugs.chromium.org/p/chromium/issues/attachment?aid=589773&signed_aid=PLB9MrSIsiAh5-VSfHCZXQ==&inline=1

Fixed: 1420656
Change-Id: I254d9fdc6f4f8896b99cdad3049af983d8265447