Enable the Trace Engine to run a subset of handlers
As we migrate to the new engine, we are executing all of it when users
use the Performance Panel. But a lot of the data is not being used yet
as the UI has not been migrated. Certain handlers can be relatively slow
to execute (particularly Renderer, Samples, and Network) as they have a
lot to process.

To mitigate the impact of running both trace engines in parallel, this
CL gives us the ability to have the model just run a subset of engines.
We can therefore run only the handlers that we need, and add to this
list as the migration continues. And then, once we are done, we can
remove most of the logic of this CL and go back to running all the
handlers again.

Note that this CL does NOT change the handlers that are run. I will do
that in a follow-up CL.

Bug: 1420931
Change-Id: I15cc48cf9cb3594248f5108c8bf9bb61d3d4f743