RPP: disable inactive handlers
This CL updates the set of active handlers in the new trace engine to
only be the ones that we currently rely on. To do this we define a new
constant in the model which lists the active handlers. To then ensure
that TypeScript knows which handlers are active (and consequently errors
if you try to use data from an inactive handler) we introduce the
PartialTraceData type, which is now used throughout the codebase.

Once all handlers are active, some of these mechanics can be removed.
But by using this type defined in the model, we ensure that we are type
safe when accessing data from the new model.

Bug: 1420931
Change-Id: I53de24f73ce4d6e0763db7be43ad9c085977d35c